+++ CHANGELOG.md
@@ -10,12 +10,14 @@
 * Check for unique names when players are provided during game random composition or creation.
 * Check for roles compatibility before game creation.
 * Games can be created by user. User can't create a game if there is already one playing.
+* Games can be canceled at any moment.
 
 ### 🛣️ Routes
 
 * Route `GET /games` added for retrieving all games. Protected with basic authentication.
 * Route `GET /games/repartition` added for getting a random team composition. Protected with basic authentication.
 * Route `POST /games` added for creating games. Protected with JWT.
+* Route `PATCH /games` added for patching games. Protected with JWT.
 * Route `GET /roles` added for getting all available roles. Protected with basic authentication.
 
 ### 💾 Database